added interactive footer

+++ src/backend.rs
@@ -22,7 +22,7 @@ pub async fn get_blog_count() -> Result<usize> {
 }
 
 #[get("/api/blog/:id")]
-pub async fn get_blog(id: i32) -> Result<String> {
+pub async fn get_blog(id: usize) -> Result<String> {
     let file_path = format!("{}/{}.md", BLOG_DIR, id);
     match fs::read_to_string(&file_path) {
         Ok(content) => Ok(content),
